refactor: split `types.ts` into multiple files in `Schema/types/*`
