+++ src/Schema/BaseSchema.test.ts
@@ -1,11 +1,11 @@
-import { Schema } from "./Schema.js";
-import type { TableKeysSchemaType, ModelSchemaType, SchemaMetadata } from "./types.js";
+import { BaseSchema } from "./BaseSchema.js";
+import type { TableKeysSchemaType, ModelSchemaType, SchemaMetadata } from "./types/index.js";
 
-describe("Schema", () => {
-  describe("Schema.validateAttributeTypes()", () => {
+describe("BaseSchema", () => {
+  describe("BaseSchema.validateAttributeTypes()", () => {
     // SHARED TEST VALUES:
 
-    /** SchemaMetadata objects (2nd param of `Schema.validateAttributeTypes`) */
+    /** SchemaMetadata objects (2nd param of `BaseSchema.validateAttributeTypes`) */
     const METADATA = {
       TK_SCHEMA: { schemaType: "TableKeysSchema" },
       M_SCHEMA: { schemaType: "ModelSchema" },
@@ -30,7 +30,7 @@ describe("Schema", () => {
       };
 
       expect(() =>
-        Schema.validateAttributeTypes(modelSchema, METADATA.M_SCHEMA)
+        BaseSchema.validateAttributeTypes(modelSchema, METADATA.M_SCHEMA)
       ).not.toThrowError();
     });
 
@@ -52,7 +52,7 @@ describe("Schema", () => {
       };
 
       expect(() =>
-        Schema.validateAttributeTypes(tableKeysSchema, METADATA.TK_SCHEMA)
+        BaseSchema.validateAttributeTypes(tableKeysSchema, METADATA.TK_SCHEMA)
       ).not.toThrowError();
     });
 
@@ -63,37 +63,39 @@ describe("Schema", () => {
       };
 
       expect(() =>
-        Schema.validateAttributeTypes(modelSchema, METADATA.M_SCHEMA)
+        BaseSchema.validateAttributeTypes(modelSchema, METADATA.M_SCHEMA)
       ).not.toThrowError();
     });
 
     test("throws a SchemaValidationError when an invalid schema is provided", () => {
-      expect(() => Schema.validateAttributeTypes([] as any, METADATA.M_SCHEMA)).toThrowError(
+      expect(() => BaseSchema.validateAttributeTypes([] as any, METADATA.M_SCHEMA)).toThrowError(
         `${ERR_MSG_PREFIX.M_SCHEMA} schema must be a plain object, but received "Array".`
       );
-      expect(() => Schema.validateAttributeTypes(null as any, METADATA.M_SCHEMA)).toThrowError(
+      expect(() => BaseSchema.validateAttributeTypes(null as any, METADATA.M_SCHEMA)).toThrowError(
         `${ERR_MSG_PREFIX.M_SCHEMA} schema must be a plain object, but received "Null".`
       );
-      expect(() => Schema.validateAttributeTypes(undefined as any, METADATA.M_SCHEMA)).toThrowError(
+      expect(() =>
+        BaseSchema.validateAttributeTypes(undefined as any, METADATA.M_SCHEMA)
+      ).toThrowError(
         `${ERR_MSG_PREFIX.M_SCHEMA} schema must be a plain object, but received "undefined".`
       );
     });
 
     test("throws a SchemaValidationError when an empty schema is provided", () => {
-      expect(() => Schema.validateAttributeTypes({} as any, METADATA.M_SCHEMA)).toThrowError(
+      expect(() => BaseSchema.validateAttributeTypes({} as any, METADATA.M_SCHEMA)).toThrowError(
         `${ERR_MSG_PREFIX.M_SCHEMA} schema does not contain any attributes.`
       );
     });
 
     test("throws a SchemaValidationError when an attribute does not specify a type", () => {
       expect(() =>
-        Schema.validateAttributeTypes({ attr: {} } as any, METADATA.M_SCHEMA)
+        BaseSchema.validateAttributeTypes({ attr: {} } as any, METADATA.M_SCHEMA)
       ).toThrowError(`${ERR_MSG_PREFIX.M_SCHEMA} attribute "attr" does not specify a "type".`);
     });
 
     test("throws a SchemaValidationError when an attribute specifies an invalid type", () => {
       expect(() =>
-        Schema.validateAttributeTypes({ attr: { type: "BAD_TYPE" } } as any, METADATA.M_SCHEMA)
+        BaseSchema.validateAttributeTypes({ attr: { type: "BAD_TYPE" } } as any, METADATA.M_SCHEMA)
       ).toThrowError(
         `${ERR_MSG_PREFIX.M_SCHEMA} attribute "attr" has an invalid "type" value (must be "string", "number", "boolean", "Buffer", "Date", "map", "array", "tuple", or "enum").`
       );
@@ -101,19 +103,25 @@ describe("Schema", () => {
 
     test(`throws a SchemaValidationError when an attribute is of type "map", "array", or "tuple", but does not specify a nested "schema"`, () => {
       expect(() =>
-        Schema.validateAttributeTypes({ mapAttr: { type: "map" } } as any, METADATA.M_SCHEMA)
+        BaseSchema.validateAttributeTypes({ mapAttr: { type: "map" } } as any, METADATA.M_SCHEMA)
       ).toThrowError(
         `${ERR_MSG_PREFIX.M_SCHEMA} attribute "mapAttr" is of type "map", but does not specify a nested "schema".`
       );
 
       expect(() =>
-        Schema.validateAttributeTypes({ arrayAttr: { type: "array" } } as any, METADATA.M_SCHEMA)
+        BaseSchema.validateAttributeTypes(
+          { arrayAttr: { type: "array" } } as any,
+          METADATA.M_SCHEMA
+        )
       ).toThrowError(
         `${ERR_MSG_PREFIX.M_SCHEMA} attribute "arrayAttr" is of type "array", but does not specify a nested "schema".`
       );
 
       expect(() =>
-        Schema.validateAttributeTypes({ tupleAttr: { type: "tuple" } } as any, METADATA.M_SCHEMA)
+        BaseSchema.validateAttributeTypes(
+          { tupleAttr: { type: "tuple" } } as any,
+          METADATA.M_SCHEMA
+        )
       ).toThrowError(
         `${ERR_MSG_PREFIX.M_SCHEMA} attribute "tupleAttr" is of type "tuple", but does not specify a nested "schema".`
       );
@@ -121,22 +129,28 @@ describe("Schema", () => {
 
     test(`throws a SchemaValidationError when an attribute is of type "map", "array", or "tuple", and "schema" is the wrong type`, () => {
       expect(() =>
-        // prettier-ignore
-        Schema.validateAttributeTypes({ attr: { type: "map", schema: [] } } as any, METADATA.M_SCHEMA)
+        BaseSchema.validateAttributeTypes(
+          { attr: { type: "map", schema: [] } } as any,
+          METADATA.M_SCHEMA
+        )
       ).toThrowError(
         `${ERR_MSG_PREFIX.M_SCHEMA} attribute "attr" is of type "map", but its nested "schema" is not an object.`
       );
 
       expect(() =>
-        // prettier-ignore
-        Schema.validateAttributeTypes({ attr: { type: "array", schema: {} } } as any, METADATA.M_SCHEMA)
+        BaseSchema.validateAttributeTypes(
+          { attr: { type: "array", schema: {} } } as any,
+          METADATA.M_SCHEMA
+        )
       ).toThrowError(
         `${ERR_MSG_PREFIX.M_SCHEMA} attribute "attr" is of type "array", but its nested "schema" is not an array.`
       );
 
       expect(() =>
-        // prettier-ignore
-        Schema.validateAttributeTypes({ attr: { type: "tuple", schema: {} } } as any, METADATA.M_SCHEMA)
+        BaseSchema.validateAttributeTypes(
+          { attr: { type: "tuple", schema: {} } } as any,
+          METADATA.M_SCHEMA
+        )
       ).toThrowError(
         `${ERR_MSG_PREFIX.M_SCHEMA} attribute "attr" is of type "tuple", but its nested "schema" is not an array.`
       );
@@ -144,15 +158,19 @@ describe("Schema", () => {
 
     test(`throws a SchemaValidationError when an attribute is of type "enum", and "oneOf" is invalid`, () => {
       expect(() =>
-        // prettier-ignore
-        Schema.validateAttributeTypes({ attr: { type: "enum", oneOf: {} } } as any, METADATA.M_SCHEMA)
+        BaseSchema.validateAttributeTypes(
+          { attr: { type: "enum", oneOf: {} } } as any,
+          METADATA.M_SCHEMA
+        )
       ).toThrowError(
         `${ERR_MSG_PREFIX.M_SCHEMA} attribute "attr" is of type "enum", but does not specify a valid "oneOf" array.`
       );
 
       expect(() =>
-        // prettier-ignore
-        Schema.validateAttributeTypes({ attr: { type: "enum", oneOf: [] } } as any, METADATA.M_SCHEMA)
+        BaseSchema.validateAttributeTypes(
+          { attr: { type: "enum", oneOf: [] } } as any,
+          METADATA.M_SCHEMA
+        )
       ).toThrowError(
         `${ERR_MSG_PREFIX.M_SCHEMA} attribute "attr" is of type "enum", but does not specify a valid "oneOf" array.`
       );
@@ -160,37 +178,52 @@ describe("Schema", () => {
 
     test(`throws a SchemaValidationError when an attribute's "default" does not align with its "type"`, () => {
       expect(() =>
-        Schema.validateAttributeTypes({ attr: { type: "string", default: 1 } }, METADATA.M_SCHEMA)
+        BaseSchema.validateAttributeTypes(
+          { attr: { type: "string", default: 1 } },
+          METADATA.M_SCHEMA
+        )
       ).toThrowError(
         `${ERR_MSG_PREFIX.M_SCHEMA} attribute "attr" specifies a "default" value of type "number", but the attribute's configured "type" is "string".`
       );
 
       expect(() =>
-        Schema.validateAttributeTypes({ attr: { type: "number", default: "" } }, METADATA.M_SCHEMA)
+        BaseSchema.validateAttributeTypes(
+          { attr: { type: "number", default: "" } },
+          METADATA.M_SCHEMA
+        )
       ).toThrowError(
         `${ERR_MSG_PREFIX.M_SCHEMA} attribute "attr" specifies a "default" value of type "string", but the attribute's configured "type" is "number".`
       );
 
       expect(() =>
-        Schema.validateAttributeTypes({ attr: { type: "boolean", default: "" } }, METADATA.M_SCHEMA)
+        BaseSchema.validateAttributeTypes(
+          { attr: { type: "boolean", default: "" } },
+          METADATA.M_SCHEMA
+        )
       ).toThrowError(
         `${ERR_MSG_PREFIX.M_SCHEMA} attribute "attr" specifies a "default" value of type "string", but the attribute's configured "type" is "boolean".`
       );
 
       expect(() =>
-        Schema.validateAttributeTypes({ attr: { type: "Buffer", default: "" } }, METADATA.M_SCHEMA)
+        BaseSchema.validateAttributeTypes(
+          { attr: { type: "Buffer", default: "" } },
+          METADATA.M_SCHEMA
+        )
       ).toThrowError(
         `${ERR_MSG_PREFIX.M_SCHEMA} attribute "attr" specifies a "default" value of type "string", but the attribute's configured "type" is "Buffer".`
       );
 
       expect(() =>
-        Schema.validateAttributeTypes({ attr: { type: "Date", default: "" } }, METADATA.M_SCHEMA)
+        BaseSchema.validateAttributeTypes(
+          { attr: { type: "Date", default: "" } },
+          METADATA.M_SCHEMA
+        )
       ).toThrowError(
         `${ERR_MSG_PREFIX.M_SCHEMA} attribute "attr" specifies a "default" value of type "string", but the attribute's configured "type" is "Date".`
       );
 
       expect(() =>
-        Schema.validateAttributeTypes(
+        BaseSchema.validateAttributeTypes(
           { attr: { type: "map", default: "", schema: { nestedAttr: { type: "string" } } } },
           METADATA.M_SCHEMA
         )
@@ -199,7 +232,7 @@ describe("Schema", () => {
       );
 
       expect(() =>
-        Schema.validateAttributeTypes(
+        BaseSchema.validateAttributeTypes(
           { attr: { type: "array", default: "", schema: [{ type: "string" }] } },
           METADATA.M_SCHEMA
         )
@@ -208,7 +241,7 @@ describe("Schema", () => {
       );
 
       expect(() =>
-        Schema.validateAttributeTypes(
+        BaseSchema.validateAttributeTypes(
           { attr: { type: "tuple", default: "", schema: [{ type: "string" }] } },
           METADATA.M_SCHEMA
         )
@@ -217,7 +250,7 @@ describe("Schema", () => {
       );
 
       expect(() =>
-        Schema.validateAttributeTypes(
+        BaseSchema.validateAttributeTypes(
           { attr: { type: "enum", default: "", oneOf: ["x"] } },
           METADATA.M_SCHEMA
         )
