+++ src/Schema/index.ts
@@ -1,4 +1,7 @@
-export * from "./Schema.js";
-export * from "./TableKeysSchema.js";
+export * from "./BaseSchema.js";
 export * from "./ModelSchema.js";
-export * from "./types.js";
+export * from "./TableKeysSchema.js";
+
+export * from "./types/index.js";
+
+export * from "./timestampAttributes.js";
