+++ src/Schema/types/AttributeName.ts
@@ -1,0 +1,4 @@
+/**
+ * An attribute name / schema key.
+ */
+export type AttributeName = string;
