+++ src/Schema/types/index.ts
@@ -1,0 +1,18 @@
+// High-level Schema types
+export * from "./TableKeysSchemaType.js";
+export * from "./ModelSchemaType.js";
+export * from "./MergeModelAndTableKeysSchema.js";
+
+// AttributeConfig base types
+export * from "./BaseAttributeConfig.js";
+export * from "./KeyAttributeConfig.js";
+export * from "./ModelSchemaAttributeConfig.js";
+
+// AttributeConfig derived types
+export * from "./AnyValidAttributeConfig.js";
+
+// Other types
+export * from "./AttributeName.js";
+export * from "./ModelSchemaEntries.js";
+export * from "./ModelSchemaOptions.js";
+export * from "./SchemaMetadata.js";
